@@ -114,9 +114,9 @@
     }
   },
   "devDependencies": {
-    "@babel/core": "^7.15.5",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.16.7",
+    "@babel/core": "^7.17.12",
+    "@babel/preset-react": "^7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
     "@electron-forge/cli": "^6.0.0-beta.63",
     "@electron-forge/maker-deb": "^6.0.0-beta.63",
     "@electron-forge/maker-dmg": "^6.0.0-beta.63",
@@ -125,26 +125,26 @@
     "@electron-forge/maker-zip": "^6.0.0-beta.63",
     "@electron-forge/plugin-webpack": "^6.0.0-beta.63",
     "@electron-forge/publisher-github": "^6.0.0-beta.63",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.6",
     "@types/d3": "^7.1.0",
     "@types/decompress": "^4.2.4",
     "@types/node-localstorage": "^1.3.0",
     "@types/os-utils": "^0.0.1",
-    "@types/react-dom": "^18.0.3",
+    "@types/react-dom": "^18.0.4",
     "@types/semver": "^7.3.9",
     "@types/uuid": "^8.3.4",
     "@types/yargs": "^17.0.10",
     "@typescript-eslint/eslint-plugin": "^5.21.0",
     "@typescript-eslint/parser": "^5.21.0",
     "@vercel/webpack-asset-relocator-loader": "^1.7.0",
-    "babel-loader": "^8.2.2",
+    "babel-loader": "^8.2.5",
     "concurrently": "^6.5.1",
     "cross-env": "^7.0.3",
     "css-loader": "^6.2.0",
-    "electron": "^18.2.0",
+    "electron": "^18.2.3",
     "electron-installer-common": "^0.10.3",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
+    "eslint": "^8.15.0",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-typescript": "^2.7.1",
@@ -158,25 +158,25 @@
     "nodemon": "^2.0.16",
     "pre-commit": "^1.2.2",
     "prettier": "^2.6.2",
-    "react-refresh": "^0.11.0",
+    "react-refresh": "^0.13.0",
     "semver-regex": ">=3.1.3",
     "style-loader": "^3.2.1",
     "typescript": "^4.6.4"
   },
   "dependencies": {
     "@babel/plugin-transform-react-jsx": "^7.17.3",
-    "@chakra-ui/icons": "^1.0.15",
-    "@chakra-ui/react": "^1.8.8",
-    "@emotion/react": "^11.8.1",
+    "@chakra-ui/icons": "^2.0.0",
+    "@chakra-ui/react": "^2.0.1",
+    "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
     "@fontsource/open-sans": "^4.5.1",
     "@react-hook/window-size": "^3.0.7",
     "@react-nano/use-event-source": "^0.11.0",
     "cross-fetch": "^3.1.5",
     "decompress": "^4.2.1",
-    "electron-log": "^4.4.1",
+    "electron-log": "^4.4.7",
     "electron-squirrel-startup": "^1.0.0",
-    "framer-motion": "^6.2.8",
+    "framer-motion": "^6.3.3",
     "html-to-image": "^1.9.0",
     "image-js": "^0.33.0",
     "node-localstorage": "^2.2.1",
@@ -184,11 +184,11 @@
     "os-utils": "^0.0.14",
     "portfinder": "^1.0.28",
     "re-resizable": "^6.9.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-flow-renderer": "^10.0.8",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-flow-renderer": "^10.2.3-next.2",
     "react-hotkeys-hook": "^3.4.4",
-    "react-icons": "^4.2.0",
+    "react-icons": "^4.3.1",
     "react-markdown": "^8.0.3",
     "scheduler": "^0.22.0",
     "semver": "^7.3.5",
