@@ -1,7 +1,7 @@
 module.exports = {
     parser: '@typescript-eslint/parser',
     extends: ['airbnb', 'plugin:react/jsx-runtime', 'plugin:prettier/recommended'],
-    plugins: ['prettier', '@typescript-eslint', 'prefer-arrow-functions'],
+    plugins: ['prettier', '@typescript-eslint'], // , 'prefer-arrow-functions'],
     globals: {
         MAIN_WINDOW_WEBPACK_ENTRY: true,
         SPLASH_SCREEN_WEBPACK_ENTRY: true,
@@ -26,15 +26,15 @@ module.exports = {
         'react/jsx-filename-extension': ['error', { extensions: ['.jsx', '.tsx'] }],
         'import/extensions': 'off',
         'react/prop-types': 'off',
-        'prefer-arrow-functions/prefer-arrow-functions': [
-            'warn',
-            {
-                classPropertiesAllowed: false,
-                disallowPrototype: false,
-                returnStyle: 'unchanged',
-                singleReturnOnly: false,
-            },
-        ],
+        // 'prefer-arrow-functions/prefer-arrow-functions': [
+        //     'warn',
+        //     {
+        //         classPropertiesAllowed: false,
+        //         disallowPrototype: false,
+        //         returnStyle: 'unchanged',
+        //         singleReturnOnly: false,
+        //     },
+        // ],
         'sort-imports': ['error', { ignoreDeclarationSort: true }],
         'import/order': [
             'error',
@@ -73,7 +73,7 @@ module.exports = {
             plugins: [
                 'prettier',
                 '@typescript-eslint',
-                'prefer-arrow-functions',
+                // 'prefer-arrow-functions',
                 'eslint-comments',
             ],
             parserOptions: {
