@@ -9,7 +9,7 @@ import {
 import { memo } from 'react';
 import useSystemUsage from '../hooks/useSystemUsage';
 
-const SystemStats = () => {
+function SystemStats() {
     const usage = useSystemUsage(2500);
     const trackColor = useColorModeValue('gray.300', 'gray.700');
     return (
@@ -74,6 +74,6 @@ const SystemStats = () => {
             )}
         </HStack>
     );
-};
+}
 
 export default memo(SystemStats);
