@@ -88,7 +88,7 @@ export const ExecutionContext = createContext<Readonly<ExecutionContextValue>>(
 );
 
 // eslint-disable-next-line @typescript-eslint/ban-types
-export const ExecutionProvider = ({ children }: React.PropsWithChildren<{}>) => {
+export function ExecutionProvider({ children }: React.PropsWithChildren<{}>) {
     const { schemata, useAnimateEdges, setIteratorPercent } = useContext(GlobalContext);
     const { useIsCpu, useIsFp16, port } = useContext(SettingsContext);
     const { sendAlert } = useContext(AlertBoxContext);
@@ -282,4 +282,4 @@ export const ExecutionProvider = ({ children }: React.PropsWithChildren<{}>) => 
             {children}
         </ExecutionContext.Provider>
     );
-};
+}
